requirements.txt
alabaster==0.7.10
args==0.1.0
Babel==2.4.0
bsub==0.3.5
chardet==2.3.0
clint==0.5.1

DateTime==4.1.1
Django==1.8.4
django-activity-stream==0.6.0
django-braces==1.9.0
django-filter==0.11.0
django-jsonfield==0.9.15
djangorestframework==3.2.3
docutils==0.13.1
ecdsa==0.13
funcsigs==0.4
imagesize==0.7.1
Jinja2==2.9.6
lxml==3.4.4
Markdown==2.6.2
MarkupSafe==1.0
mock==1.3.0
MySQL-python==1.2.5
nose==1.3.7
paramiko==1.15.2
pbr==1.7.0
Pinax==0.9a2
pinax-eventlog==1.0.0
pycrypto==2.6.1
Pygments==2.2.0
python-dateutil==2.4.2
pytz==2016.4
requests==2.7.0
rt==1.0.9
selenium==2.53.6
sh==1.11
six==1.9.0
snowballstemmer==1.2.1
Sphinx==1.6.1
sphinxcontrib-websupport==1.0.1
typing==3.6.1
zope.interface==4.2.0
#cx-Oracle==5.2.1